drivers/linuxgpiodv2: introduce new driver for libgpiod v2 API
This driver is a copy of the existing linuxgpiod driver but adapted
to the newer v2 API of libgpiod.

I've chosen to add it in parallel to the existing driver so that it
is possible to have releases with support for both.

I decided to make a dedicated new module to prevent having a lot of
ifdef-ery inside the C file itself - in my eyes, the small amount of
additionl makefile magic is simpler.

The version of libgpiod is detected during configure time based on
the pkg-config version detection since the library itself does not
provide any defines or similar stuff.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>