drivers/linuxgpiodv2: introduce new driver for libgpiod v2 API
This driver is a copy of the existing linuxgpiod driver but adapted
to the newer v2 API of libgpiod.

I've chosen to add it in parallel to the existing driver so that it
is possible to have releases with support for both - however, since
you usuall can only have a single libgpiod version installed on your
system, it will result in compile errors if you do not disable the
non-matching one.

Signed-off-by: Michael Heimpold <mhei@heimpold.de>